requirements.txt
africastalking==1.1.9
appdirs==1.4.3
apturl==0.5.2
asgiref==3.3.1
astroid==2.3.3
bcrypt==3.1.7
blinker==1.4
Brlapi==0.7.0
CacheControl==0.12.6
cairocffi==1.2.0
CairoSVG==2.5.1
certifi==2019.11.28
cffi==1.14.5
chardet==3.0.4
click==7.1.2
colorama==0.4.3
command-not-found==0.3
contextlib2==0.5.5
crypto==1.4.1
cryptography==2.8
cssselect2==0.4.1
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
defusedxml==0.6.0
distlib==0.3.0
distro==1.4.0
distro-info===0.23ubuntu1
dj-database-url==0.4.2
Django==3.1.5
django-ckeditor==6.0.0
django-cors-headers==3.2.1
django-countries==6.1.3
django-countries-plus==1.3.1
django-filter==2.2.0
django-js-asset==1.2.2
django-restframework==0.0.1
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
duplicity==0.8.12.0
entrypoints==0.3
fasteners==0.14.1
filelock==3.0.12
freeze==3.0
future==0.18.2
gunicorn==20.0.4
html5lib==1.0.1
httplib2==0.14.0
idna==2.8
ipaddr==2.2.0
isort==4.3.21
jovian==0.2.14
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-object-proxy==1.4.3
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.2.1
MarkupSafe==1.1.0
mccabe==0.6.1
monotonic==1.5
msgpack==0.6.2
Naked==0.1.31
netifaces==0.10.4
numpy==1.18.1
oauthlib==3.1.0
olefile==0.46
opencv-contrib-python==4.1.2.30
packaging==20.3
paramiko==2.6.0
pbr==5.5.1
pep517==0.8.2
pexpect==4.6.0
Pillow==7.0.0
progress==1.5
protobuf==3.6.1
pycairo==1.16.2
pycparser==2.20
pycryptodome==3.9.4
pycups==1.9.73
PyGObject==3.36.0
PyJWT==1.7.1
pylint==2.4.4
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==2.4.6
Pyphen==0.10.0
pypng==0.0.20
PyQRCode==1.2.1
pyRFC3339==1.1
python-apt==2.0.0+ubuntu0.20.4.3
python-barcode==0.11.0
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
pytoml==0.1.21
pytz==2020.5
pyxdg==0.26
PyYAML==5.3
pyzbar==0.1.8
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
retrying==1.3.3
schema==0.7.1
SecretStorage==2.3.1
shellescape==3.4.1
simplejson==3.16.0
six==1.14.0
sqlparse==0.4.1
systemd-python==234
testresources==2.0.1
tinycss2==1.1.0
toml==0.10.1
ubuntu-advantage-tools==20.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.7
usb-creator==0.3.7
uuid==1.30
virtualenv==16.7.9
virtualenvwrapper-win==1.2.5
wadllib==1.3.3
WeasyPrint==52.2
webencodings==0.5.1
wrapt==1.11.2
xkit==0.0.0
xlwt==1.3.0
